Refactor SA-CASSCF response into free functions
