+++ forte2/mcopt/mc_optimizer.py
@@ -2,7 +2,6 @@ from dataclasses import dataclass, field
 from typing import ClassVar
 
 import numpy as np
-import scipy.sparse.linalg as spla
 from numpy.typing import NDArray
 
 
@@ -19,7 +18,6 @@ from forte2.orbitals import (
     validate_final_orbitals,
 )
 from forte2.jkbuilder import RestrictedMOIntegrals, SpinorbitalIntegrals
-from forte2.lib.ci_helpers import CISigmaBuilder
 from forte2.helpers import logger, LBFGS
 from forte2.system.basis_utils import BasisInfo
 from forte2.system import ModelSystem
@@ -331,12 +329,6 @@ class MCOptimizerBase(Method):
         self.E_ci = np.array(self.ci_solver.E)
         self.E_avg = self.ci_solver.compute_average_energy()
         self.E = self.E_avg
-        # Keep the orbital response base point synchronized with the final CI
-        # vectors. The last CI diagonalization may slightly change the RDMs
-        # after the final orbital microiteration.
-        self.g1_act = self.make_average_1rdm()
-        g2_act = self.make_average_2rdm()
-        self.orb_opt.set_rdms(self.g1_act, g2_act)
         logger.log_info1(
             f"{'Final CI':>10} {self.E_avg:>20.10f} {self.E_orb:>20.10f} {'-':>12} {'-':>12} {'-':>12} {'-':>8} {'':>8}"
         )
@@ -574,1112 +566,6 @@ class MCOptimizerBase(Method):
 
 
 class MCOptimizer(MCOptimizerBase):
-    def _get_ci_response_layout(self):
-        r"""Construct the root-major coefficient layout and its dimension.
-
-        If root :math:`\alpha` contains :math:`n_\alpha` CSF coefficients,
-        the flattened response vector represented by the returned slices is
-
-        .. math::
-
-            \mathbf x
-            =\mathbf x_0\oplus\mathbf x_1\oplus\cdots
-             \oplus\mathbf x_{d-1},
-            \qquad
-            n_{\mathrm{CI}}=\sum_{\alpha=0}^{d-1}n_\alpha.
-        """
-        layout = []
-        start = 0
-        for absolute_root, (state_index, root_in_state) in enumerate(
-            self.ci_solver.sa_info.absolute_root_map
-        ):
-            sub_solver = self.ci_solver.sub_solvers[state_index]
-            stop = start + sub_solver.basis_size
-            layout.append(
-                (
-                    absolute_root,
-                    state_index,
-                    root_in_state,
-                    slice(start, stop),
-                )
-            )
-            start = stop
-        return tuple(layout), start
-
-    def get_ci_response_layout(self):
-        r"""Return the root-major layout of the flattened CI response vector.
-
-        The coefficient ordering :math:`\mathbb L=((\alpha,M))_J` is defined
-        by
-
-        .. math::
-
-            \mathbf x
-            =\bigoplus_{\alpha=0}^{d-1}\mathbf x_\alpha,
-            \qquad
-            x_J=(x_\alpha)_M
-            \quad\text{for }J\leftrightarrow(\alpha,M)\in\mathbb L.
-
-        Returns
-        -------
-        tuple[tuple[int, int, int, slice], ...]
-            One entry per absolute root. Each entry contains
-            (absolute_root, state_index, root_in_state, coefficient_slice).
-            Coefficients inside each slice use that sub-solver's CSF ordering.
-        """
-        layout, _ = self._get_ci_response_layout()
-        return layout
-
-    def _validate_response_root(self, root):
-        if isinstance(root, bool) or not isinstance(root, (int, np.integer)):
-            raise TypeError("The target response root must be an integer.")
-        nroots = len(self.ci_solver.sa_info.absolute_root_map)
-        if root < 0 or root >= nroots:
-            raise ValueError(
-                f"Expected a target response root in [0, {nroots}), got {root}."
-            )
-        return int(root)
-
-    def _validate_ci_response_vector(self, ci_vector):
-        layout, nci = self._get_ci_response_layout()
-        ci_vector = np.asarray(ci_vector)
-        if ci_vector.shape != (nci,):
-            raise ValueError(
-                f"Expected a CI response vector with shape ({nci},), "
-                f"got {ci_vector.shape}."
-            )
-        if np.iscomplexobj(ci_vector):
-            raise TypeError(
-                "The nonrelativistic orbital--CI response vector must be real."
-            )
-        return ci_vector.astype(float, copy=False), layout
-
-    def _project_ci_response_vector(self, ci_vector, layout):
-        r"""Project each root block out of its solved-state CI subspace.
-
-        For every state solver :math:`s`, this applies
-
-        .. math::
-
-            \mathbf Q_s\mathbf x_\alpha
-            =\left(\mathbf I_s-
-              \sum_{\gamma\in\mathcal R_s}
-              \mathbf c_\gamma\mathbf c_\gamma^{\mathsf T}\right)
-              \mathbf x_\alpha,
-            \qquad \alpha\in\mathcal R_s.
-        """
-        projected = np.empty_like(ci_vector)
-        for _, state_index, _, coefficient_slice in layout:
-            sub_solver = self.ci_solver.sub_solvers[state_index]
-            solved_roots = sub_solver.evecs[:, : sub_solver.nroot]
-            root_vector = ci_vector[coefficient_slice]
-            projected[coefficient_slice] = root_vector - solved_roots @ (
-                solved_roots.T @ root_vector
-            )
-        return projected
-
-    def project_ci_response_vector(self, ci_vector):
-        r"""Project a root-major vector into the CI response space.
-
-        In every absolute-root block, this method applies the projector for
-        that root's state solver,
-
-        .. math::
-
-            \mathbf Q_s
-            =\mathbf I_s-\sum_{\gamma\in\mathcal R_s}
-             \mathbf c_\gamma\mathbf c_\gamma^{\mathsf T},
-
-        where ``R_s`` contains all solved roots represented in the same CSF
-        space.  The projection removes CI normalization directions and
-        rotations within the solved-root subspace.
-
-        Parameters
-        ----------
-        ci_vector : np.ndarray
-            Root-major flattened real vector with shape ``(nci,)``.
-
-        Returns
-        -------
-        np.ndarray
-            Projected root-major vector with shape ``(nci,)``.
-        """
-        ci_vector, layout = self._validate_ci_response_vector(ci_vector)
-        return self._project_ci_response_vector(ci_vector, layout)
-
-    def _compute_ci_response_rdms(self, ci_vector, layout):
-        r"""Form root-summed overlap and transition-RDM responses.
-
-        For the root-major vector :math:`\mathbf x`, this returns
-
-        .. math::
-
-            s[\mathbf x]
-            &=\sum_{\alpha=0}^{d-1}\left(
-              S^{\mathbf x_\alpha,\mathbf c_\alpha}
-             +S^{\mathbf c_\alpha,\mathbf x_\alpha}\right),\\
-            \gamma^{u}_{v}[\mathbf x]
-            &=\sum_{\alpha=0}^{d-1}\left[
-              (\gamma^{u}_{v})^{\mathbf x_\alpha,\mathbf c_\alpha}
-             +(\gamma^{u}_{v})^{\mathbf c_\alpha,\mathbf x_\alpha}\right],\\
-            \gamma^{uv}_{wx}[\mathbf x]
-            &=\sum_{\alpha=0}^{d-1}\left[
-              (\gamma^{uv}_{wx})^{\mathbf x_\alpha,\mathbf c_\alpha}
-             +(\gamma^{uv}_{wx})^{\mathbf c_\alpha,\mathbf x_\alpha}\right].
-
-        The implementation evaluates both bra and ket transition densities
-        explicitly in the determinant basis.
-        """
-        nact = self.mo_space.nactv
-        overlap_response = 0.0
-        g1_response = np.zeros((nact,) * 2, dtype=float)
-        g2_response = np.zeros((nact,) * 4, dtype=float)
-
-        for _, state_index, root_in_state, coefficient_slice in layout:
-            response = ci_vector[coefficient_slice]
-            if not np.any(response):
-                continue
-
-            sub_solver = self.ci_solver.sub_solvers[state_index]
-            reference = sub_solver.evecs[:, root_in_state]
-            overlap_response += np.dot(response, reference)
-            overlap_response += np.dot(reference, response)
-
-            response_det = sub_solver.csf_C_to_det_C(response)
-            reference_det = sub_solver.csf_C_to_det_C(reference)
-            sigma_builder = sub_solver.ci_sigma_builder
-            g1_response += sigma_builder.sf_1rdm(response_det, reference_det)
-            g1_response += sigma_builder.sf_1rdm(reference_det, response_det)
-            g2_response += sigma_builder.sf_2rdm(response_det, reference_det)
-            g2_response += sigma_builder.sf_2rdm(reference_det, response_det)
-
-        return overlap_response, g1_response, g2_response
-
-    def compute_orbital_ci_hessian_vector_product(self, ci_vector):
-        r"""Apply the CI contribution to the orbital response equation.
-
-        The input is the concatenation of one real CSF vector for each absolute
-        root, in the layout returned by get_ci_response_layout. For root
-        alpha, its segment x_alpha is contracted with the reference CI vector
-        c_alpha to form the bra-plus-ket transition overlap and RDMs. The
-        returned vector is
-
-        .. math::
-
-            [\mathcal A^{\mathrm{oc}}\mathbf x]_I
-            =
-            2\left[
-              (A^{\mathrm{oc}}[\mathbf x])^{q_I}_{p_I}
-             -(A^{\mathrm{oc}}[\mathbf x])^{p_I}_{q_I}
-            \right].
-
-        No state-average weights multiply x_alpha in this block because it is
-        the orbital derivative of the unweighted CI residual Lagrangian. For a
-        direct derivative of the state-averaged orbital gradient, the
-        equivalent physical CI displacement is x_alpha / w_alpha.
-
-        Parameters
-        ----------
-        ci_vector : np.ndarray
-            Root-major flattened real CI multiplier vector.
-
-        Returns
-        -------
-        np.ndarray
-            Orbital response vector with shape (nrot,) and nrr C-order.
-        """
-        ci_vector, layout = self._validate_ci_response_vector(ci_vector)
-        responses = self._compute_ci_response_rdms(ci_vector, layout)
-        intermediates = self.orb_opt._build_ci_orbital_response_intermediates()
-        return self.orb_opt._compute_ci_orbital_response_from_rdms(
-            *responses, intermediates
-        )
-
-    def compute_orbital_ci_hessian(self):
-        r"""Build the full rectangular orbital--CI response block.
-
-        Column J is the orbital response obtained by applying
-        compute_orbital_ci_hessian_vector_product to CI unit vector J. Columns
-        use the root-major layout returned by get_ci_response_layout:
-
-        .. math::
-
-            (\mathcal A^{\mathrm{oc}})_{IJ}
-            =\left[\mathcal A^{\mathrm{oc}}\mathbf e_J\right]_I,
-            \qquad
-            \mathcal A^{\mathrm{oc}}\in
-            \mathbb R^{n_{\mathrm{rot}}\times n_{\mathrm{CI}}}.
-
-        Returns
-        -------
-        np.ndarray
-            The orbital--CI block with shape (nrot, nci).
-        """
-        layout, nci = self._get_ci_response_layout()
-        intermediates = self.orb_opt._build_ci_orbital_response_intermediates()
-        hessian = np.empty((self.orb_opt.nrot, nci), dtype=float)
-        unit = np.zeros(nci, dtype=float)
-        for column in range(nci):
-            unit[column] = 1.0
-            responses = self._compute_ci_response_rdms(unit, layout)
-            hessian[:, column] = self.orb_opt._compute_ci_orbital_response_from_rdms(
-                *responses, intermediates
-            )
-            unit[column] = 0.0
-        return hessian
-
-    def _compute_ci_orbital_hessian_vector_product(
-        self, orbital_vector, layout, intermediates
-    ):
-        r"""Apply the CI--orbital block with precomputed base-point tensors.
-
-        The orbital optimizer first produces the response-integral triple
-        ``(Ecore_response, H_response, V_response)`` defining
-        :math:`\hat H[\mathbf z]`.  One temporary :class:`CISigmaBuilder` is
-        constructed for each distinct state sub-solver.  For every absolute
-        root, the kernel converts its reference CSF vector to determinants,
-        applies the response Hamiltonian, converts the sigma vector back to
-        CSFs, and stores
-
-        .. math::
-
-            2w_\alpha\hat H[\mathbf z]\mathbf c_\alpha
-
-        in that root's coefficient slice.
-
-        This private routine assumes that ``orbital_vector`` is a validated
-        real ``(nrot,)`` array, ``layout`` is the current root-major layout,
-        and ``intermediates`` was built at the current orbitals.  It applies no
-        coefficient-space projector.
-
-        Returns
-        -------
-        np.ndarray
-            Root-major raw CSF response with shape ``(nci,)``.
-        """
-        scalar_response, one_body_response, two_body_response = (
-            self.orb_opt._compute_active_space_hamiltonian_response(
-                orbital_vector, intermediates
-            )
-        )
-        response = np.empty(layout[-1][-1].stop, dtype=float)
-        builders = {}
-
-        for absolute_root, state_index, root_in_state, coefficient_slice in layout:
-            sub_solver = self.ci_solver.sub_solvers[state_index]
-            if state_index not in builders:
-                builder = CISigmaBuilder(
-                    sub_solver.ci_strings,
-                    scalar_response,
-                    one_body_response,
-                    two_body_response,
-                    sub_solver.log_level,
-                )
-                builder.set_memory(sub_solver.ci_params.ci_builder_memory)
-                algorithm = sub_solver.ci_params.ci_algorithm.lower()
-                builder.set_algorithm("kh" if algorithm == "exact" else algorithm)
-                builders[state_index] = builder
-
-            reference = sub_solver.evecs[:, root_in_state]
-            sigma_csf = self._apply_ci_hamiltonian_to_csf(
-                sub_solver, builders[state_index], reference
-            )
-            weight = self.ci_solver.weights_flat[absolute_root]
-            response[coefficient_slice] = 2.0 * weight * sigma_csf
-
-        return response
-
-    @staticmethod
-    def _apply_ci_hamiltonian_to_csf(sub_solver, sigma_builder, vector):
-        r"""Apply a determinant-basis Hamiltonian to a real CSF vector.
-
-        If :math:`T` maps CSF coefficients to determinant coefficients, this
-        helper returns the CSF representation of the sigma vector
-
-        .. math::
-
-            \boldsymbol\sigma_{\mathrm{CSF}}
-            =T^{\mathsf T}\mathbf H_{\mathrm{det}}T\mathbf c_{\mathrm{CSF}}
-            \equiv\mathbf H_{\mathrm{CSF}}\mathbf c_{\mathrm{CSF}}.
-        """
-        vector_det = sub_solver.csf_C_to_det_C(vector)
-        sigma_det = np.empty(sub_solver.ndet, dtype=float)
-        sigma_builder.Hamiltonian(vector_det, sigma_det)
-        sigma_csf = np.empty(sub_solver.basis_size, dtype=float)
-        sub_solver.spin_adapter.det_C_to_csf_C(sigma_det, sigma_csf)
-        return sigma_csf
-
-    def compute_ci_orbital_hessian_vector_product(self, orbital_vector):
-        r"""Apply the orbital contribution to the CI response equation.
-
-        For an orbital direction ``z`` in the nonredundant pair ordering, the
-        active-space scalar, one-electron, and two-electron integrals are
-        differentiated to define the response Hamiltonian ``H[z]``.  The
-        root-major CSF result is
-
-        .. math::
-
-            [\mathcal A^{\mathrm{co}}\mathbf z]_\alpha
-            =2w_\alpha\hat H[\mathbf z]\mathbf c_\alpha.
-
-        The state-average weight ``w_alpha`` is included once.  The returned
-        vector is the raw coefficient-space action; no normalization or
-        state-rotation projector is applied.
-
-        Parameters
-        ----------
-        orbital_vector : np.ndarray
-            Real orbital-rotation direction with shape ``(nrot,)`` and the
-            same pair ordering as the orbital optimizer.
-
-        Returns
-        -------
-        np.ndarray
-            Root-major flattened CSF response vector with shape ``(nci,)``.
-        """
-        orbital_vector = self.orb_opt._validate_orbital_response_vector(orbital_vector)
-        layout, _ = self._get_ci_response_layout()
-        intermediates = self.orb_opt._build_orbital_ci_response_intermediates()
-        return self._compute_ci_orbital_hessian_vector_product(
-            orbital_vector, layout, intermediates
-        )
-
-    def compute_ci_orbital_hessian(self):
-        r"""Build the full rectangular CI--orbital response block.
-
-        Column ``I`` is obtained by applying
-        :meth:`compute_ci_orbital_hessian_vector_product` to orbital unit
-        vector ``I``.  Rows use the root-major CSF layout returned by
-        :meth:`get_ci_response_layout`.  With ``W_C`` repeating each root's SA
-        weight over its CSF block, the real mixed derivatives obey
-
-        .. math::
-
-            \mathcal A^{\mathrm{co}}
-            =W_{\mathrm C}(\mathcal A^{\mathrm{oc}})^{\mathsf T}.
-
-        Returns
-        -------
-        np.ndarray
-            The CI--orbital block with shape ``(nci, nrot)``.
-        """
-        layout, nci = self._get_ci_response_layout()
-        intermediates = self.orb_opt._build_orbital_ci_response_intermediates()
-        hessian = np.empty((nci, self.orb_opt.nrot), dtype=float)
-        unit = np.zeros(self.orb_opt.nrot, dtype=float)
-        for column in range(self.orb_opt.nrot):
-            unit[column] = 1.0
-            hessian[:, column] = self._compute_ci_orbital_hessian_vector_product(
-                unit, layout, intermediates
-            )
-            unit[column] = 0.0
-        return hessian
-
-    def _compute_ci_ci_hessian_vector_product(self, ci_vector, layout):
-        r"""Apply the raw CI--CI block to a validated coefficient vector.
-
-        For every absolute root, the corresponding final-state CI sigma
-        builder evaluates ``H @ x_alpha``.  The root's scalar energy shift is
-        then subtracted and the real-coordinate factor of two is applied:
-
-        .. math::
-
-            [\mathcal A^{\mathrm{cc}}\mathbf x]_\alpha
-            =2(\mathbf H_\alpha-E_\alpha\mathbf I_\alpha)
-             \mathbf x_\alpha.
-
-        Different absolute-root slices do not couple.  This private kernel
-        assumes a validated vector and current layout and applies neither SA
-        weights nor normalization/state-rotation projectors.
-
-        Returns
-        -------
-        np.ndarray
-            Root-major raw CSF response with shape ``(nci,)``.
-        """
-        response = np.empty_like(ci_vector)
-        for absolute_root, state_index, _, coefficient_slice in layout:
-            sub_solver = self.ci_solver.sub_solvers[state_index]
-            root_vector = ci_vector[coefficient_slice]
-            sigma = self._apply_ci_hamiltonian_to_csf(
-                sub_solver, sub_solver.ci_sigma_builder, root_vector
-            )
-            response[coefficient_slice] = 2.0 * (
-                sigma - self.E_ci[absolute_root] * root_vector
-            )
-        return response
-
-    def compute_ci_ci_hessian_vector_product(self, ci_vector):
-        r"""Apply the nonrelativistic CI--CI response block.
-
-        The input and output concatenate one real CSF vector per absolute
-        root in the layout returned by :meth:`get_ci_response_layout`.  For
-        root ``alpha``, this method computes
-
-        .. math::
-
-            [\mathcal A^{\mathrm{cc}}\mathbf x]_\alpha
-            =2(\mathbf H_\alpha-E_\alpha\mathbf I_\alpha)
-             \mathbf x_\alpha.
-
-        This is the raw real-coordinate block.  It contains no SA weights and
-        no coefficient-space projector, so each reference CI vector is a null
-        direction of its root block.
-
-        Parameters
-        ----------
-        ci_vector : np.ndarray
-            Root-major flattened real CI response vector with shape
-            ``(nci,)``.
-
-        Returns
-        -------
-        np.ndarray
-            Root-major flattened CI response with shape ``(nci,)``.
-        """
-        ci_vector, layout = self._validate_ci_response_vector(ci_vector)
-        return self._compute_ci_ci_hessian_vector_product(ci_vector, layout)
-
-    def compute_ci_ci_hessian(self):
-        r"""Build the full root-major CI--CI response matrix.
-
-        Column ``J`` is obtained by applying
-        :meth:`compute_ci_ci_hessian_vector_product` to coefficient-space unit
-        vector ``J``.  Equivalently,
-
-        .. math::
-
-            (\mathcal A^{\mathrm{cc}})_{JK}
-            =\left[\mathcal A^{\mathrm{cc}}\mathbf e_K\right]_J,
-            \qquad
-            \mathcal A^{\mathrm{cc}}
-            =\bigoplus_{\alpha=0}^{d-1}
-             2(\mathbf H_\alpha-E_\alpha\mathbf I_\alpha).
-
-        Returns
-        -------
-        np.ndarray
-            Dense raw CI--CI response matrix with shape ``(nci, nci)``.
-        """
-        layout, nci = self._get_ci_response_layout()
-        hessian = np.empty((nci, nci), dtype=float)
-        unit = np.zeros(nci, dtype=float)
-        for column in range(nci):
-            unit[column] = 1.0
-            hessian[:, column] = self._compute_ci_ci_hessian_vector_product(
-                unit, layout
-            )
-            unit[column] = 0.0
-        return hessian
-
-    def compute_ci_response_vector_product(self, orbital_vector, ci_vector):
-        r"""Apply the CI row of the coupled CASSCF response matrix.
-
-        .. math::
-
-            \mathbf y^{\mathrm c}
-            =\mathcal A^{\mathrm{co}}\mathbf z
-             +\mathcal A^{\mathrm{cc}}\mathbf x.
-
-        The result is raw and unprojected in coefficient space.
-
-        Parameters
-        ----------
-        orbital_vector : np.ndarray
-            Real nonredundant orbital direction with shape ``(nrot,)``.
-        ci_vector : np.ndarray
-            Root-major flattened real CI direction with shape ``(nci,)``.
-
-        Returns
-        -------
-        np.ndarray
-            Combined root-major CI response with shape ``(nci,)``.
-        """
-        orbital_vector = self.orb_opt._validate_orbital_response_vector(orbital_vector)
-        ci_vector, layout = self._validate_ci_response_vector(ci_vector)
-        intermediates = self.orb_opt._build_orbital_ci_response_intermediates()
-        orbital_response = self._compute_ci_orbital_hessian_vector_product(
-            orbital_vector, layout, intermediates
-        )
-        ci_response = self._compute_ci_ci_hessian_vector_product(ci_vector, layout)
-        return orbital_response + ci_response
-
-    def compute_orbital_response_b_vector(self, root):
-        r"""Build the target-state orbital ``b`` vector.
-
-        For absolute root ``alpha``, this method evaluates the positive
-        target-energy orbital gradient at the converged state-averaged
-        orbitals,
-
-        .. math::
-
-            (\mathbf b^{\mathrm o}_\alpha)_I
-            =(g^\alpha_{\mathrm{F2}})_I
-            =2\left[
-              (A^\alpha)^{q_I}_{p_I}-(A^\alpha)^{p_I}_{q_I}
-            \right].
-
-        The root-specific spin-free 1- and 2-RDMs are used without an SA
-        weight.  The result is the ``b`` vector itself, not the signed linear
-        solver right-hand side: the coupled orbital equation is
-
-        .. math::
-
-            (\mathcal A^{\mathrm{oo}})^{\mathsf T}\mathbf z
-            +\mathcal A^{\mathrm{oc}}\mathbf x
-            =-\mathbf b^{\mathrm o}_\alpha.
-
-        Parameters
-        ----------
-        root : int
-            Absolute target-root index in state-average ordering.
-
-        Returns
-        -------
-        np.ndarray
-            Real target-state orbital gradient with shape ``(nrot,)`` and the
-            orbital optimizer's nonredundant-pair ordering.
-        """
-        root = self._validate_response_root(root)
-        g1 = self.make_sf_1rdm(root)
-        g2 = self.make_sf_2rdm(root)
-        intermediates = self.orb_opt._build_ci_orbital_response_intermediates()
-        return self.orb_opt._compute_ci_orbital_response_from_rdms(
-            1.0, g1, g2, intermediates
-        )
-
-    def _compute_raw_ci_response_b_vector(self, root, layout):
-        r"""Build the unprojected real target-energy CI gradient.
-
-        For target absolute root :math:`\alpha`, this returns
-
-        .. math::
-
-            (\widetilde{\mathbf b}^{\mathrm c}_\alpha)_\beta
-            =2\delta_{\alpha\beta}\mathbf H_\beta\mathbf c_\beta.
-
-        The coefficient-space projector is deliberately not applied here.
-        """
-        _, state_index, root_in_state, coefficient_slice = layout[root]
-        sub_solver = self.ci_solver.sub_solvers[state_index]
-        reference = sub_solver.evecs[:, root_in_state]
-        sigma = self._apply_ci_hamiltonian_to_csf(
-            sub_solver, sub_solver.ci_sigma_builder, reference
-        )
-        raw_b = np.zeros(layout[-1][-1].stop, dtype=float)
-        raw_b[coefficient_slice] = 2.0 * sigma
-        return raw_b
-
-    def compute_ci_response_b_vector(self, root):
-        r"""Build the projected target-state CI ``b`` vector.
-
-        For target absolute root ``alpha``, the unprojected positive
-        target-energy gradient has only one nonzero root block,
-
-        .. math::
-
-            (\widetilde{\mathbf b}^{\mathrm c}_\alpha)_\beta
-            =2\delta_{\alpha\beta}\mathbf H_\beta\mathbf c_\beta.
-
-        This method forms that vector with the final-state CI sigma builder
-        and applies :meth:`project_ci_response_vector`.  Because a converged
-        target CI vector obeys ``H c_alpha = E_alpha c_alpha``, its gradient is
-        entirely in the removed solved-root subspace and
-        ``b_ci`` is zero to the CI convergence tolerance.
-
-        The returned vector is the positive projected ``b`` vector, not the
-        signed solver right-hand side.  No SA weight is included.  A zero CI
-        ``b`` vector does not imply a zero CI response multiplier because the
-        orbital--CI block couples it to the nonzero orbital right-hand side.
-
-        Parameters
-        ----------
-        root : int
-            Absolute target-root index in state-average ordering.
-
-        Returns
-        -------
-        np.ndarray
-            Projected root-major CI ``b`` vector with shape ``(nci,)``.
-        """
-        root = self._validate_response_root(root)
-        layout, _ = self._get_ci_response_layout()
-        raw_b = self._compute_raw_ci_response_b_vector(root, layout)
-        return self._project_ci_response_vector(raw_b, layout)
-
-    def compute_projected_response_vector_product(self, orbital_vector, ci_vector):
-        r"""Apply the gauge-fixed projected coupled response operator.
-
-        If ``Q`` is the block-diagonal CI response projector and ``P = I-Q``,
-        this method returns
-
-        .. math::
-
-            \begin{pmatrix}
-             (\mathcal A^{\mathrm{oo}})^{\mathsf T}\mathbf z
-             +\mathcal A^{\mathrm{oc}}Q\mathbf x\\
-             Q\left(\mathcal A^{\mathrm{co}}\mathbf z
-             +\mathcal A^{\mathrm{cc}}Q\mathbf x\right)+P\mathbf x
-            \end{pmatrix}.
-
-        The ``P @ x`` term fixes the otherwise singular removed CI directions
-        to zero and leaves the physical projected equations unchanged.
-
-        Parameters
-        ----------
-        orbital_vector : np.ndarray
-            Real nonredundant orbital vector with shape ``(nrot,)``.
-        ci_vector : np.ndarray
-            Root-major real CI vector with shape ``(nci,)``.
-
-        Returns
-        -------
-        tuple[np.ndarray, np.ndarray]
-            Orbital and gauge-fixed projected CI products.
-        """
-        orbital_vector = self.orb_opt._validate_orbital_response_vector(orbital_vector)
-        ci_vector, layout = self._validate_ci_response_vector(ci_vector)
-        intermediates = self.orb_opt._build_coupled_response_intermediates()
-        return self._compute_projected_response_vector_product(
-            orbital_vector, ci_vector, layout, intermediates
-        )
-
-    def _compute_projected_response_vector_product(
-        self, orbital_vector, ci_vector, layout, intermediates
-    ):
-        r"""Apply the projected response operator using a shared workspace.
-
-        With :math:`\mathcal P_{\mathrm C}=\mathbf I_{\mathrm C}-
-        \mathcal Q_{\mathrm C}`, the returned pair is
-
-        .. math::
-
-            \begin{pmatrix}\mathbf y^{\mathrm o}\\\mathbf y^{\mathrm c}\end{pmatrix}
-            =
-            \begin{pmatrix}
-             (\mathcal A^{\mathrm{oo}})^{\mathsf T}
-             &\mathcal A^{\mathrm{oc}}\mathcal Q_{\mathrm C}\\
-             \mathcal Q_{\mathrm C}\mathcal A^{\mathrm{co}}
-             &\mathcal Q_{\mathrm C}\mathcal A^{\mathrm{cc}}
-              \mathcal Q_{\mathrm C}+\mathcal P_{\mathrm C}
-            \end{pmatrix}
-            \begin{pmatrix}\mathbf z\\\mathbf x\end{pmatrix}.
-        """
-        projected_ci = self._project_ci_response_vector(ci_vector, layout)
-
-        orbital_intermediates, density_intermediates, hamiltonian_intermediates = (
-            intermediates
-        )
-        orbital_product = (
-            self.orb_opt._compute_orbital_hessian_transpose_vector_product(
-                orbital_vector, orbital_intermediates
-            )
-        )
-        density_response = self._compute_ci_response_rdms(projected_ci, layout)
-        orbital_product += self.orb_opt._compute_ci_orbital_response_from_rdms(
-            *density_response, density_intermediates
-        )
-
-        raw_ci_product = self._compute_ci_orbital_hessian_vector_product(
-            orbital_vector, layout, hamiltonian_intermediates
-        )
-        raw_ci_product += self._compute_ci_ci_hessian_vector_product(
-            projected_ci, layout
-        )
-        ci_product = self._project_ci_response_vector(raw_ci_product, layout)
-        ci_product += ci_vector - projected_ci
-        return orbital_product, ci_product
-
-    @staticmethod
-    def _invert_response_diagonal(diagonal):
-        r"""Return a sign-preserving, safely regularized inverse diagonal.
-
-        For :math:`s=\max(1,\lVert\mathbf d\rVert_\infty)` and
-        :math:`\epsilon=10^{-6}`, the returned entries are
-
-        .. math::
-
-            m_i&=\widetilde d_i^{-1},\\
-            \widetilde d_i&=
-            \begin{cases}
-              d_i,&|d_i|\ge\epsilon s,\\
-              \operatorname{sgn}_{+}(d_i)\epsilon s,&|d_i|<\epsilon s,
-            \end{cases}
-
-        where :math:`\operatorname{sgn}_{+}(0)=1`.
-        """
-        diagonal = np.asarray(diagonal, dtype=float)
-        scale = max(1.0, float(np.max(np.abs(diagonal), initial=0.0)))
-        floor = 1.0e-6 * scale
-        signs = np.where(diagonal < 0.0, -1.0, 1.0)
-        regularized = np.where(np.abs(diagonal) < floor, signs * floor, diagonal)
-        return 1.0 / regularized
-
-    def _build_response_preconditioner(self, layout, nrot, nci):
-        r"""Build a matrix-free block-Jacobi inverse for coupled GMRES.
-
-        Let :math:`\mathbf d^{\mathrm o}` be the orbital optimizer's diagonal
-        Hessian model and let
-
-        .. math::
-
-            (d^{\mathrm c}_\alpha)_M
-            =2\left[(H_\alpha)_{MM}-E_\alpha\right].
-
-        After applying the regularized inverse defined by
-        :meth:`_invert_response_diagonal`, this routine represents
-
-        .. math::
-
-            \mathcal M^{-1}
-            \begin{pmatrix}\mathbf v^{\mathrm o}\\\mathbf v^{\mathrm c}\end{pmatrix}
-            =
-            \begin{pmatrix}
-              (\widetilde D^{\mathrm o})^{-1}\mathbf v^{\mathrm o}\\
-              \mathcal Q_{\mathrm C}(\widetilde D^{\mathrm c})^{-1}
-              \mathcal Q_{\mathrm C}\mathbf v^{\mathrm c}
-              +\mathcal P_{\mathrm C}\mathbf v^{\mathrm c}
-            \end{pmatrix}.
-
-        It stores only the two inverse diagonals and applies the CI projectors
-        matrix-free.
-        """
-        orbital_diagonal = self.orb_opt._mat_to_vec(self.orb_opt._compute_orbhess())
-        orbital_inverse = self._invert_response_diagonal(orbital_diagonal)
-
-        ci_inverse = np.empty(nci, dtype=float)
-        state_diagonals = {}
-        for absolute_root, state_index, _, coefficient_slice in layout:
-            sub_solver = self.ci_solver.sub_solvers[state_index]
-            if state_index not in state_diagonals:
-                state_diagonals[state_index] = (
-                    sub_solver.ci_sigma_builder.form_Hdiag_csf(
-                        sub_solver.dets,
-                        sub_solver.spin_adapter,
-                        False,
-                    )
-                )
-            diagonal = 2.0 * (state_diagonals[state_index] - self.E_ci[absolute_root])
-            ci_inverse[coefficient_slice] = self._invert_response_diagonal(diagonal)
-
-        dimension = nrot + nci
-
-        def matvec(vector):
-            r"""Apply the block-Jacobi inverse to one coupled vector.
-
-            .. math::
-
-                \mathbf y=\mathcal M^{-1}\mathbf v.
-            """
-            product = np.empty(dimension, dtype=float)
-            product[:nrot] = orbital_inverse * vector[:nrot]
-            ci_vector = vector[nrot:]
-            projected_ci = self._project_ci_response_vector(ci_vector, layout)
-            product[nrot:] = self._project_ci_response_vector(
-                ci_inverse * projected_ci, layout
-            )
-            product[nrot:] += ci_vector - projected_ci
-            return product
-
-        return spla.LinearOperator((dimension, dimension), matvec=matvec, dtype=float)
-
-    def solve_state_specific_response(
-        self,
-        root,
-        *,
-        r_tol=1.0e-10,
-        maxiter=None,
-    ):
-        r"""Solve the projected coupled response equations for one target root.
-
-        Preconditioned GMRES is applied to the matrix-free gauge-fixed system
-
-        .. math::
-
-            \begin{pmatrix}
-             (\mathcal A^{\mathrm{oo}})^{\mathsf T}
-             & \mathcal A^{\mathrm{oc}}Q\\
-             Q\mathcal A^{\mathrm{co}} &
-             Q\mathcal A^{\mathrm{cc}}Q+P
-            \end{pmatrix}
-            \begin{pmatrix}\mathbf z_\alpha\\\mathbf x_\alpha\end{pmatrix}
-            =-
-            \begin{pmatrix}
-             \mathbf b^{\mathrm o}_\alpha\\
-             \mathbf b^{\mathrm c}_\alpha
-            \end{pmatrix},
-
-        with ``P = I - Q``.  A matrix-free block-Jacobi preconditioner uses the
-        orbital optimizer's diagonal Hessian and the diagonal CI Hamiltonian;
-        neither a dense response block nor a CI-to-RDM Jacobian is built.  The
-        CI solution is projected once more before it is returned to remove
-        roundoff in the gauge components.
-
-        Parameters
-        ----------
-        root : int
-            Absolute target-root index in state-average ordering.
-        r_tol : float, optional
-            Relative GMRES residual tolerance.
-        maxiter : int or None, optional
-            Maximum number of GMRES restart cycles.  The SciPy default is used
-            when omitted.
-
-        Returns
-        -------
-        tuple[np.ndarray, np.ndarray]
-            The orbital response ``z_alpha`` with shape ``(nrot,)`` and the
-            projected root-major CI response ``x_alpha`` with shape
-            ``(nci,)``.
-
-        Raises
-        ------
-        RuntimeError
-            If GMRES does not converge.
-        """
-        root = self._validate_response_root(root)
-        if not np.isscalar(r_tol) or r_tol <= 0.0:
-            raise ValueError(f"r_tol must be positive, got {r_tol}.")
-        if maxiter is not None and (
-            isinstance(maxiter, bool)
-            or not isinstance(maxiter, (int, np.integer))
-            or maxiter < 1
-        ):
-            raise ValueError(f"maxiter must be a positive integer, got {maxiter}.")
-
-        layout, nci = self._get_ci_response_layout()
-        nrot = self.orb_opt.nrot
-        dimension = nrot + nci
-        intermediates = self.orb_opt._build_coupled_response_intermediates()
-        _, density_intermediates, _ = intermediates
-        orbital_b = self.orb_opt._compute_ci_orbital_response_from_rdms(
-            1.0,
-            self.make_sf_1rdm(root),
-            self.make_sf_2rdm(root),
-            density_intermediates,
-        )
-        raw_ci_b = self._compute_raw_ci_response_b_vector(root, layout)
-        ci_b = self._project_ci_response_vector(raw_ci_b, layout)
-        rhs = -np.concatenate((orbital_b, ci_b))
-
-        def matvec(vector):
-            r"""Apply the gauge-fixed coupled operator :math:`\mathscr A`.
-
-            .. math::
-
-                \mathbf y=\mathscr A\mathbf v,
-
-            where :math:`\mathscr A` is the block matrix displayed in this
-            method's response equation.
-            """
-            orbital_product, ci_product = (
-                self._compute_projected_response_vector_product(
-                    vector[:nrot], vector[nrot:], layout, intermediates
-                )
-            )
-            product = np.empty(dimension, dtype=float)
-            product[:nrot] = orbital_product
-            product[nrot:] = ci_product
-            return product
-
-        operator = spla.LinearOperator(
-            (dimension, dimension), matvec=matvec, dtype=float
-        )
-        preconditioner = self._build_response_preconditioner(layout, nrot, nci)
-        restart = min(dimension, 50)
-        solution, info = spla.gmres(
-            operator,
-            rhs,
-            rtol=float(r_tol),
-            atol=0.0,
-            restart=restart,
-            maxiter=maxiter,
-            M=preconditioner,
-        )
-        if info != 0:
-            if info > 0:
-                reason = f"did not converge after {info} iterations"
-            else:
-                reason = f"failed with status {info}"
-            raise RuntimeError(f"The coupled CASSCF response solve {reason}.")
-
-        orbital_response = solution[:nrot]
-        ci_response = self._project_ci_response_vector(solution[nrot:], layout)
-        return orbital_response, ci_response
-
-    def solve_orbital_response_vector(self, root, *, r_tol=1.0e-10, maxiter=None):
-        r"""Solve and return only the target-root orbital response vector.
-
-        If :math:`\mathscr A` denotes the gauge-fixed block matrix in
-        :meth:`solve_state_specific_response`, this method returns
-
-        .. math::
-
-            \mathbf z_\alpha
-            =\begin{pmatrix}\mathbf I_{\mathrm o}&\mathbf 0\end{pmatrix}
-             \mathscr A^{-1}
-             \left[-\begin{pmatrix}
-              \mathbf b^{\mathrm o}_\alpha\\
-              \mathbf b^{\mathrm c}_\alpha
-             \end{pmatrix}\right].
-        """
-        orbital_response, _ = self.solve_state_specific_response(
-            root, r_tol=r_tol, maxiter=maxiter
-        )
-        return orbital_response
-
-    def compute_omega(
-        self,
-        root,
-        orbital_response=None,
-        ci_response=None,
-        *,
-        r_tol=1.0e-10,
-        maxiter=None,
-    ):
-        r"""Compute the relaxed MO orthogonality multiplier for one root.
-
-        For target root :math:`\alpha`, let :math:`A^\alpha` be its orbital
-        Lagrangian, :math:`A^{\mathrm{oc}}[\mathbf x_\alpha]` the contribution
-        from the root-major CI multiplier, and :math:`\dot{\bar A}[\mathbf
-        z_\alpha]` the directional response of the state-averaged orbital
-        Lagrangian.  In the current real nonrelativistic convention this method
-        forms
-
-        .. math::
-
-            \Omega_\alpha
-            = A^\alpha
-            + A^{\mathrm{oc}}[\mathbf x_\alpha]
-            + \dot{\bar A}[\mathbf z_\alpha]
-            + Z_\alpha\bar A-\bar A Z_\alpha,
-
-            \omega_\alpha
-            = \frac{1}{2}
-              (\Omega_\alpha+\Omega_\alpha^{\mathsf T}).
-
-        The commutator is required because the orbital-pair generators form a
-        moving frame: differentiating ``z.T @ g`` with respect to the full MO
-        coefficients is not just the directional derivative of ``A``.  The
-        returned matrix is in the current MO basis.  Its AO counterpart for
-        the overlap-derivative gradient term is ``C @ omega @ C.T``.
-
-        If neither response vector is supplied, the coupled projected response
-        equations are solved first.  If response vectors are supplied, both
-        must be given; the CI vector is projected into the physical response
-        space before its transition RDMs are formed.
-
-        Parameters
-        ----------
-        root : int
-            Absolute target-root index in state-average ordering.
-        orbital_response : np.ndarray or None, optional
-            Solved nonredundant orbital multiplier ``z_alpha``.
-        ci_response : np.ndarray or None, optional
-            Solved root-major CI multiplier ``x_alpha``.
-        r_tol : float, optional
-            Relative GMRES tolerance used only when solving the response.
-        maxiter : int or None, optional
-            Maximum GMRES restart cycles used only when solving the response.
-
-        Returns
-        -------
-        np.ndarray
-            Real symmetric ``omega_alpha`` with shape ``(nmo, nmo)`` in the
-            current MO basis.
-        """
-        root = self._validate_response_root(root)
-        supplied_orbital = orbital_response is not None
-        supplied_ci = ci_response is not None
-        if supplied_orbital != supplied_ci:
-            raise ValueError(
-                "orbital_response and ci_response must either both be supplied "
-                "or both be omitted."
-            )
-
-        if not supplied_orbital:
-            orbital_response, ci_response = self.solve_state_specific_response(
-                root, r_tol=r_tol, maxiter=maxiter
-            )
-        else:
-            orbital_response = self.orb_opt._validate_orbital_response_vector(
-                orbital_response
-            )
-            ci_response, layout = self._validate_ci_response_vector(ci_response)
-            ci_response = self._project_ci_response_vector(ci_response, layout)
-
-        layout, _ = self._get_ci_response_layout()
-        (
-            orbital_intermediates,
-            density_intermediates,
-            _,
-        ) = self.orb_opt._build_coupled_response_intermediates()
-
-        target_A = self.orb_opt._build_orbital_lagrangian_from_rdms(
-            1.0,
-            self.make_sf_1rdm(root),
-            self.make_sf_2rdm(root),
-            density_intermediates,
-        )
-        ci_A = self.orb_opt._build_orbital_lagrangian_from_rdms(
-            *self._compute_ci_response_rdms(ci_response, layout),
-            density_intermediates,
-        )
-        average_A = self.orb_opt._build_orbital_lagrangian_from_rdms(
-            1.0,
-            self.make_average_1rdm(),
-            self.make_average_2rdm(),
-            density_intermediates,
-        )
-
-        orbital_A = self.orb_opt._compute_orbital_lagrangian_response(
-            orbital_response, orbital_intermediates
-        )
-        Z = self.orb_opt._vec_to_mat(orbital_response)
-        orbital_A += Z @ average_A - average_A @ Z
-
-        Omega = target_A + ci_A + orbital_A
-        return 0.5 * (Omega + Omega.T)
-
-    def compute_orbital_response_vector_product(self, orbital_vector, ci_vector):
-        r"""Apply the orbital row of the coupled CASSCF response matrix.
-
-        .. math::
-
-            \mathbf y^{\mathrm o}
-            =
-            (\mathcal A^{\mathrm{oo}})^{\mathsf T}\mathbf z
-            +
-            \mathcal A^{\mathrm{oc}}\mathbf x.
-
-        Parameters
-        ----------
-        orbital_vector : np.ndarray
-            Real nonredundant orbital direction with shape (nrot,).
-        ci_vector : np.ndarray
-            Root-major flattened real CI multiplier vector.
-
-        Returns
-        -------
-        np.ndarray
-            Combined orbital response with shape (nrot,).
-        """
-        orbital_vector = self.orb_opt._validate_orbital_response_vector(orbital_vector)
-        ci_vector, layout = self._validate_ci_response_vector(ci_vector)
-        orbital_intermediates, density_intermediates, _ = (
-            self.orb_opt._build_coupled_response_intermediates()
-        )
-        density_response = self._compute_ci_response_rdms(ci_vector, layout)
-        ci_response = self.orb_opt._compute_ci_orbital_response_from_rdms(
-            *density_response, density_intermediates
-        )
-        orbital_response = (
-            self.orb_opt._compute_orbital_hessian_transpose_vector_product(
-                orbital_vector, orbital_intermediates
-            )
-        )
-        return orbital_response + ci_response
-
     def make_sd_1rdm(
         self,
         left_root: int,
@@ -1721,16 +607,12 @@ class MCOptimizer(MCOptimizerBase):
 
         This implementation supports real nonrelativistic and complex
         two-component state-specific CASSCF/GASSCF wave functions, including
-        SF- and SO-X2C-1e Hamiltonians.  It also supports an individual root of
-        a real nonrelativistic SA-CASSCF/GASSCF wave function, including roots
-        from different spin-adapted CI state solvers and constrained
-        optimizations with frozen inter-GAS rotations, when
-        ``final_orbitals='original'``.  ``root`` must be specified for a state
-        average; it defaults to zero only for a single-root calculation.
-        Frozen-core and frozen-virtual response and active-frozen rotations
-        are not supported. Point and Gaussian nuclear charge distributions are
-        supported; Gaussian charges require libcint. Requesting any unsupported
-        feature raises ``NotImplementedError``.
+        SF- and SO-X2C-1e Hamiltonians. It also supports individual roots of
+        real nonrelativistic SA-CASSCF/GASSCF wave functions when
+        ``final_orbitals='original'``. Frozen-core and frozen-virtual response
+        and active-frozen rotations are not supported. Point and Gaussian
+        nuclear charge distributions are supported; Gaussian charges require
+        libcint. Requesting any unsupported feature raises ``NotImplementedError``.
         Both the orbital optimization and all CI roots must be converged; an
         unconverged wave function raises ``RuntimeError`` because the
         stationary-gradient expression does not apply.
@@ -1739,20 +621,17 @@ class MCOptimizer(MCOptimizerBase):
         and UHF gradients:
 
         .. math::
+            E^x =
+            E_\mathrm{NN}^x
+            + h^x_{\mu\nu}\Gamma_{\mu\nu}
+            - S^x_{\mu\nu} W^S_{\mu\nu}
+            + W^P_{\mu\nu}(P|\mu\nu)^x
+            + W_{PQ}(P|Q)^x.
 
-            E_\alpha^x
-            =V_\mathrm{NN}^x
-             +\sum_{\mu\nu}D^{\mathrm{rel},\alpha}_{\mu\nu}h^x_{\mu\nu}
-             -\sum_{\mu\nu}W^{S,\alpha}_{\mu\nu}S^x_{\mu\nu}
-             +\sum_{P\mu\nu}W^{P,\alpha}_{\mu\nu}(P|\mu\nu)^x
-             +\sum_{PQ}W^\alpha_{PQ}(P|Q)^x.
-
-        Here :math:`D^{\mathrm{rel},\alpha}` is the full spin-free
-        one-particle density; for a state-specific calculation it reduces to
-        the stationary density.  :math:`W^{S,\alpha}=C\omega_\alpha
-        C^{\mathsf T}` is the AO energy-weighted density, and
-        :math:`W^{P,\alpha}_{\mu\nu}` and :math:`W^\alpha_{PQ}` are the
-        density-fitted two-electron derivative weights defined in
+        Here :math:`\Gamma_{\mu\nu}` is the full spin-free one-particle
+        density, :math:`W^S_{\mu\nu}` is the AO energy-weighted density, and
+        :math:`W^P_{\mu\nu}` and :math:`W_{PQ}` are the density-fitted
+        two-electron derivative weights defined in
         ``docs/technical_notes/df_gradients.tex``.
 
         Parameters
