+++ forte2/mcopt/orbital_optimizer.py
@@ -44,26 +44,8 @@ class OrbOptimizer:
 
     def set_rdms(self, g1, g2):
         self.g1 = g1
-        self.g2 = self._make_working_2rdm(g2)
-
-    @staticmethod
-    def _make_working_2rdm(g2):
-        r"""Convert a spin-free 2-RDM to the orbital-optimizer convention.
-
-        In the notation of the technical note, the returned tensor is
-
-        .. math::
-
-            \bar D_{tu,vw}
-            =\frac{1}{2}\left(
-              \bar\gamma^{tv}_{uw}+\bar\gamma^{uv}_{tw}
-            \right).
-
-        The same linear permutation is used for state-specific and transition
-        2-RDMs, with the bars replaced by the corresponding density label.
-        """
         # '2RDM' defined as in [eq (6)]
-        return 0.5 * (np.einsum("prqs->pqrs", g2) + np.einsum("qrps->pqrs", g2))
+        self.g2 = 0.5 * (np.einsum("prqs->pqrs", g2) + np.einsum("qrps->pqrs", g2))
 
     def get_active_space_ints(self):
         """
@@ -90,145 +72,6 @@ class OrbOptimizer:
         hess = self._compute_orbhess()
         h0 = self._mat_to_vec(hess)
         return h0
-
-    def compute_orbital_hessian_vector_product(self, vector):
-        r"""Apply the nonrelativistic orbital--orbital response matrix.
-
-        Let :math:`\mathbb K=((p_I,q_I))_{I=0}^{n_\mathrm{rot}-1}` be the
-        ordered orbital pairs for which ``nrr[p_I, q_I]`` is true.  Their order
-        is exactly NumPy's C-order boolean-indexing order.  The input vector is
-        embedded in an antisymmetric generator
-
-        .. math::
-
-            Z_{pq}(\mathbf z)
-            = \sum_I z_I\left(
-                \delta_{p p_I}\delta_{q q_I}
-                - \delta_{p q_I}\delta_{q p_I}
-              \right),
-
-        and the method returns
-
-        .. math::
-
-            y_I
-            = \sum_J (\mathcal A^{\mathrm{oo}})_{IJ}z_J
-            = \left.\frac{d}{d\epsilon}
-              (\bar g_{\mathrm{F2}})^{q_I}_{p_I}
-              \left(Ce^{\epsilon Z(\mathbf z)};
-                    \bar\gamma,\bar D\right)
-              \right|_{\epsilon=0},
-
-        where :math:`(\bar g_{\mathrm{F2}})^q_p=
-        2(\bar A^q_p-\bar A^p_q)` is Forte2's nonrelativistic orbital
-        gradient.  Equivalently,
-        :math:`(\mathcal A^{\mathrm{oo}})_{IJ}` is the derivative obtained by
-        setting :math:`\mathbf z=\mathbf e_J` in the equation above.
-
-        This definition assumes real, orthonormal restricted spatial orbitals
-        and a mask containing exactly one orientation of every retained
-        unordered pair.  It is a local derivative at the current ``C``.  The
-        stored active 1-RDM ``g1`` and working 2-RDM ``g2`` (normally the
-        state-averaged RDMs), AO Hamiltonian, DF integrals, and nuclear geometry
-        are held fixed; only the orbital-dependent densities, Fock matrices,
-        and MO integral transformations respond.  Thus this is the
-        orbital--orbital partial Hessian, with neither CI nor nuclear response
-        included.  The derivative is of the unscreened analytic gradient
-        expression; the numerical ``1e-12`` screening in :meth:`gradient` is
-        not differentiated.
-
-        Parameters
-        ----------
-        vector : np.ndarray
-            Real nonredundant orbital-rotation vector with shape ``(nrot,)``.
-
-        Returns
-        -------
-        np.ndarray
-            Orbital Hessian applied to ``vector``, in the same ordered-pair
-            basis and with shape ``(nrot,)``.
-        """
-        vector = self._validate_orbital_response_vector(vector)
-        intermediates = self._build_orbital_response_intermediates()
-        return self._compute_orbital_hessian_vector_product(vector, intermediates)
-
-    def compute_orbital_hessian_transpose_vector_product(self, vector):
-        r"""Apply the transpose fixed-RDM orbital response matrix.
-
-        Let
-
-        .. math::
-
-            H_{IJ}=D_J\bar g_I
-
-        be the forward response defined by
-        :meth:`compute_orbital_hessian_vector_product`.  For the
-        antisymmetric generator :math:`Z(\mathbf z)` and the unsymmetrized
-        state-averaged orbital Lagrangian :math:`\bar A`, this method returns
-
-        .. math::
-
-            [H^{\mathsf T}\mathbf z]_I
-            =2\left(B^{q_I}_{p_I}-B^{p_I}_{q_I}\right),
-            \qquad
-            B=\dot{\bar A}[\mathbf z]+Z\bar A-\bar A Z.
-
-        The commutator is the moving-frame correction.  It has no
-        antisymmetric contribution when all nonredundant state-average
-        orbital gradients vanish, in which case :math:`H^{\mathsf T}=H`.
-        It must be retained for a constrained optimization with frozen
-        orbital rotations because the omitted gradient components may be
-        nonzero.  The CASSCF adjoint equations use this transpose action.
-
-        Parameters
-        ----------
-        vector : np.ndarray
-            Real nonredundant orbital multiplier with shape ``(nrot,)``.
-
-        Returns
-        -------
-        np.ndarray
-            :math:`H^{\mathsf T}\mathbf z` in the ordered-pair basis.
-        """
-        vector = self._validate_orbital_response_vector(vector)
-        intermediates = self._build_orbital_response_intermediates()
-        A_h = self._build_orbital_lagrangian_from_response_intermediates(*intermediates)
-        return self._compute_orbital_hessian_transpose_vector_product(
-            vector, (*intermediates, A_h)
-        )
-
-    def compute_orbital_hessian(self):
-        r"""Build the nonrelativistic orbital--orbital response matrix.
-
-        Column :math:`j` is obtained by applying the same matrix-free response
-        kernel used by :meth:`compute_orbital_hessian_vector_product` to unit
-        vector :math:`j`.  Explicitly,
-
-        .. math::
-
-            (\mathcal A^{\mathrm{oo}})_{IJ}
-            =\left[\mathcal A^{\mathrm{oo}}\mathbf e_J\right]_I,
-            \qquad
-            \mathcal A^{\mathrm{oo}}\in
-            \mathbb R^{n_{\mathrm{rot}}\times n_{\mathrm{rot}}}.
-
-        The active RDMs are held fixed.
-
-        Returns
-        -------
-        np.ndarray
-            Dense orbital Hessian in the nonredundant rotation basis.
-        """
-        intermediates = self._build_orbital_response_intermediates()
-        hessian = np.empty((self.nrot, self.nrot), dtype=float)
-        unit = np.zeros(self.nrot, dtype=float)
-        for column in range(self.nrot):
-            unit[column] = 1.0
-            hessian[:, column] = self._compute_orbital_hessian_vector_product(
-                unit, intermediates
-            )
-            unit[column] = 0.0
-        return hessian
 
     def compute_orbital_lagrangian(self):
         r"""
@@ -287,743 +130,6 @@ class OrbOptimizer:
 
     def _mat_to_vec(self, R):
         return R[self.nrr]
-
-    def _validate_orbital_response_vector(self, vector):
-        vector = np.asarray(vector)
-        if vector.shape != (self.nrot,):
-            raise ValueError(
-                f"Expected an orbital-response vector with shape ({self.nrot},), "
-                f"got {vector.shape}."
-            )
-        if np.iscomplexobj(vector):
-            raise TypeError("The nonrelativistic orbital-response vector must be real.")
-        return vector.astype(float, copy=False)
-
-    def _build_orbital_lagrangian_from_response_intermediates(
-        self, Fcore_mo, Fact_mo, B_ga
-    ):
-        r"""Build the unsymmetrized base-point orbital Lagrangian.
-
-        From the fixed-RDM response intermediates, this evaluates
-
-        .. math::
-
-            \bar A^m_p
-            &=2(F_{\mathrm C}+\bar F_{\mathrm A})^m_p,
-            &&m\in\mathbb C,\\
-            \bar A^u_p
-            &=\sum_v(F_{\mathrm C})^v_p\bar\gamma^v_u
-              +\sum_{Ptvw}B^P_{pt}B^P_{vw}\bar D_{tu,vw},
-            &&u\in\mathbb A,\\
-            \bar A^e_p&=0,
-            &&e\in\mathbb V.
-
-        Here ``g1`` is :math:`\bar\gamma`, ``g2`` is the working density
-        :math:`\bar D`, and ``B_ga`` is :math:`B^P_{pu}`.  The matrix is kept
-        unsymmetrized because its omitted frozen-rotation gradient components
-        enter the transpose orbital response through
-        :math:`Z\bar A-\bar A Z`.
-
-        Returns
-        -------
-        np.ndarray
-            The nonzero hole columns :math:`\bar A^h_p`, with
-            :math:`h\in\mathbb C\cup\mathbb A`, and shape ``(nmo, nhole)``.
-        """
-        A_h = np.empty((self.C.shape[1], self.ncore + self.nact), dtype=float)
-        A_h[:, : self.ncore] = 2.0 * (Fcore_mo + Fact_mo)[:, self.core]
-        A_h[:, self.ncore :] = np.einsum(
-            "rv,vu->ru", Fcore_mo[:, self.actv], self.g1, optimize=True
-        )
-        B_aa = B_ga[:, self.actv, :]
-        contracted_density = np.einsum("Pvw,tuvw->Ptu", B_aa, self.g2, optimize=True)
-        A_h[:, self.ncore :] += np.einsum(
-            "Prt,Ptu->ru", B_ga, contracted_density, optimize=True
-        )
-        return A_h
-
-    def _build_orbital_response_intermediates(self):
-        r"""Build the current-orbital tensors shared by response applications.
-
-        For real restricted orbitals, define the AO core and active densities
-
-        .. math::
-
-            P_{\mathrm{C}}=C_{\mathrm{C}}C_{\mathrm{C}}^{\mathsf T},\qquad
-            \bar P_{\mathrm{A}}
-            =C_{\mathrm{A}}\bar\gamma C_{\mathrm{A}}^{\mathsf T}.
-
-        This method returns the three tensors
-
-        .. math::
-
-            F_{\mathrm{C}}
-            &=C^{\mathsf T}\left(h+2J[P_{\mathrm{C}}]-K[P_{\mathrm{C}}]\right)C,\\
-            \bar F_{\mathrm{A}}
-            &=C^{\mathsf T}\left(J[\bar P_{\mathrm{A}}]
-              -\tfrac12K[\bar P_{\mathrm{A}}]\right)C,\\
-            B^P_{pu}
-            &=\sum_{\mu\nu}C_{\mu p}B^P_{\mu\nu}C_{\nu u}.
-
-        Here ``g1`` is :math:`\bar\gamma`, :math:`B^P_{\mu\nu}` is the
-        orthonormalized DF three-index tensor, and the AO maps are defined by
-
-        .. math::
-
-            J[D]_{\mu\nu}
-            &=\sum_{P\rho\sigma}B^P_{\mu\nu}B^P_{\rho\sigma}D_{\sigma\rho},\\
-            K[D]_{\mu\nu}
-            &=\sum_{P\rho\sigma}B^P_{\mu\sigma}D_{\sigma\rho}B^P_{\nu\rho}.
-
-        The first two returned arrays are in the full current MO basis; the
-        third retains only one general and one active MO index.  This is the
-        largest transformed DF block required by the matrix-free Hessian
-        action; a full ``(naux, nmo, nmo)`` tensor is never formed.
-
-        These tensors contain no response themselves.  They are fixed base-
-        point quantities that may be reused for multiple directions or all
-        columns of the dense Hessian.  They are valid only while ``C``,
-        ``Ccore``, ``Cact``, ``g1``, ``hcore``, and the DF tensor remain
-        unchanged.  The working 2-RDM ``g2`` is not needed at this stage.
-
-        Returns
-        -------
-        tuple[np.ndarray, np.ndarray, np.ndarray]
-            ``(Fcore_mo, Fact_mo, B_ga)`` with shapes ``(nmo, nmo)``,
-            ``(nmo, nmo)``, and ``(naux, nmo, nact)``.
-
-        """
-        Fcore_ao = self.fock_builder.build_core_fock(self.Ccore, hcore=self.hcore)
-        Fact_ao = self.fock_builder.build_active_fock(self.Cact, self.g1)
-        Fcore_mo = self._transform_ao_operator(Fcore_ao, self.C)
-        Fact_mo = self._transform_ao_operator(Fact_ao, self.C)
-        B_ga = self._transform_df_block(self.C, self.Cact)
-        return Fcore_mo, Fact_mo, B_ga
-
-    def _build_coupled_response_intermediates(self):
-        r"""Build one shared, active-index DF workspace for a response solve.
-
-        Returns the orbital--orbital, orbital--CI, and CI--orbital tuples used
-        by their respective private matrix-vector kernels.  Arrays common to
-        several blocks are shared by reference.  In particular, only
-        ``B[P,p,u]`` and its active-active view are retained, so the workspace
-        contains no full transformed ``B[P,p,q]`` tensor and no generalized
-        four-index integral block.
-
-        In the notation of the technical note, the returned object is
-
-        .. math::
-
-            \mathcal I_{\mathrm{coupled}}
-            =\left(
-              (F_{\mathrm C},\bar F_{\mathrm A},B^P_{pu},\bar A^h_p),
-              (F_{\mathrm C},B^P_{pu}),
-              (F_{\mathrm C}^{\mathrm{AO}},F_{\mathrm C},B^P_{pu})
-             \right),
-
-        where
-
-        .. math::
-
-            B^P_{pu}=\sum_{\mu\nu}
-            C_{\mu p}B^P_{\mu\nu}C_{\nu u}.
-
-        The three entries are respectively the intermediates for
-        :math:`\mathcal A^{\mathrm{oo}}`, :math:`\mathcal A^{\mathrm{oc}}`,
-        and :math:`\mathcal A^{\mathrm{co}}`.
-        """
-        Fcore_ao = self.fock_builder.build_core_fock(self.Ccore, hcore=self.hcore)
-        Fact_ao = self.fock_builder.build_active_fock(self.Cact, self.g1)
-        Fcore_mo = self._transform_ao_operator(Fcore_ao, self.C)
-        Fact_mo = self._transform_ao_operator(Fact_ao, self.C)
-        B_ga = self._transform_df_block(self.C, self.Cact)
-        A_h = self._build_orbital_lagrangian_from_response_intermediates(
-            Fcore_mo, Fact_mo, B_ga
-        )
-        return (
-            (Fcore_mo, Fact_mo, B_ga, A_h),
-            (Fcore_mo, B_ga),
-            (Fcore_ao, Fcore_mo, B_ga),
-        )
-
-    def _transform_df_block(self, left, right):
-        r"""Transform the AO DF tensor with two rectangular factors.
-
-        For input coefficient matrices :math:`L` and :math:`R`, this returns
-
-        .. math::
-
-            B^P_{ab}[L,R]
-            =\sum_{\mu\nu}L_{\mu a}B^P_{\mu\nu}R_{\nu b}.
-        """
-        return np.einsum(
-            "Pmn,mp,nq->Ppq",
-            self.fock_builder.B_Pmn,
-            left,
-            right,
-            optimize=True,
-        )
-
-    def _build_transition_fock_response(
-        self,
-        response_orbitals,
-        reference_orbitals,
-        metric=None,
-        *,
-        coulomb_factor,
-        exchange_factor,
-    ):
-        r"""Build a Fock response from a symmetric low-rank transition density.
-
-        Let :math:`C_L` and :math:`C_R` denote ``response_orbitals`` and
-        ``reference_orbitals``.  If a real symmetric ``metric`` :math:`G` is
-        supplied, the symmetrized metric
-        :math:`G_s=(G+G^{\mathsf T})/2` and its signed eigendecomposition
-        define
-
-        .. math::
-
-            G_s&=U\Lambda U^{\mathsf T},\\
-            L&=C_LU|\Lambda|^{1/2},\\
-            R&=C_RU\operatorname{sgn}(\Lambda)|\Lambda|^{1/2}.
-
-        With no metric, :math:`L=C_L` and :math:`R=C_R`.  The density and
-        returned AO Fock response are
-
-        .. math::
-
-            D&=LR^{\mathsf T}+RL^{\mathsf T},\\
-            \dot F^{\mathrm{AO}}[D]
-             &=c_JJ[D]-c_KK[D].
-
-        The exchange contraction is evaluated from the half transforms
-
-        .. math::
-
-            Y^{P,L}_{\mu a}&=\sum_\nu B^P_{\mu\nu}L_{\nu a},\\
-            Y^{P,R}_{\mu a}&=\sum_\nu B^P_{\mu\nu}R_{\nu a},\\
-            K[D]_{\mu\nu}
-             &=\sum_{Pa}\left(
-               Y^{P,L}_{\mu a}Y^{P,R}_{\nu a}
-              +Y^{P,R}_{\mu a}Y^{P,L}_{\nu a}\right),
-
-        so no three-general-index response tensor is formed.  The signed
-        factorization permits indefinite CI transition densities.
-        """
-        if metric is None:
-            left = response_orbitals
-            right = reference_orbitals
-        else:
-            eigenvalues, eigenvectors = np.linalg.eigh(0.5 * (metric + metric.T))
-            roots = np.sqrt(np.abs(eigenvalues))
-            left = response_orbitals @ (eigenvectors * roots[None, :])
-            right = reference_orbitals @ (
-                eigenvectors * (np.sign(eigenvalues) * roots)[None, :]
-            )
-
-        density_response = left @ right.T + right @ left.T
-        J_response = self.fock_builder.build_J([density_response])[0]
-
-        if self.fock_builder.store_B_nPm:
-            left_half = self.fock_builder.B_nPm @ left
-            right_half = self.fock_builder.B_nPm @ right
-            K_lr = np.tensordot(left_half, right_half, axes=([1, 2], [1, 2]))
-        else:
-            B = self.fock_builder.B_Pmn
-            left_half = np.einsum("Pms,si->Pmi", B, left, optimize=True)
-            right_half = np.einsum("Pms,si->Pmi", B, right, optimize=True)
-            K_lr = np.einsum("Pmi,Pni->mn", left_half, right_half, optimize=True)
-        K_response = K_lr + K_lr.T
-        return coulomb_factor * J_response - exchange_factor * K_response
-
-    def _compute_orbital_hessian_vector_product(self, vector, intermediates):
-        r"""Evaluate one fixed-RDM orbital Hessian--vector product.
-
-        Pair ordering and the embedding :math:`\mathbf z\mapsto Z(\mathbf z)`
-        are defined by :meth:`compute_orbital_hessian_vector_product`.  Starting
-        from :math:`\dot C=CZ`, this kernel evaluates
-
-        .. math::
-
-            \dot P_{\mathrm{C}}
-            &=\dot C_{\mathrm{C}}C_{\mathrm{C}}^{\mathsf T}
-              +C_{\mathrm{C}}\dot C_{\mathrm{C}}^{\mathsf T},\\
-            \dot{\bar P}_{\mathrm{A}}
-            &=\dot C_{\mathrm{A}}\bar\gamma C_{\mathrm{A}}^{\mathsf T}
-              +C_{\mathrm{A}}\bar\gamma\dot C_{\mathrm{A}}^{\mathsf T},\\
-            \dot F_{\mathrm{C}}^{\mathrm{AO}}
-            &=2J[\dot P_{\mathrm{C}}]-K[\dot P_{\mathrm{C}}],\\
-            \dot{\bar F}_{\mathrm{A}}^{\mathrm{AO}}
-            &=J[\dot{\bar P}_{\mathrm{A}}]
-              -\tfrac12K[\dot{\bar P}_{\mathrm{A}}].
-
-        For :math:`X\in\{\mathrm C,\mathrm A\}`, the corresponding MO-basis
-        response and the transformed DF-factor response are
-
-        .. math::
-
-            \dot F_X
-            &=Z^{\mathsf T} F_X+F_XZ
-              +C^{\mathsf T}\dot F_X^{\mathrm{AO}}C,\\
-            \dot B^P_{pq}
-            &=\sum_s\left(Z_{sp}B^P_{sq}+B^P_{ps}Z_{sq}\right).
-
-        With :math:`\langle pv|tw\rangle=\sum_PB^P_{pt}B^P_{vw}`, the
-        integral response is
-
-        .. math::
-
-            \dot{\langle pv|tw\rangle}
-            =\sum_P\left(\dot B^P_{pt}B^P_{vw}
-                         +B^P_{pt}\dot B^P_{vw}\right).
-
-        Holding ``g1`` (:math:`\bar\gamma`) and the internal working 2-RDM
-        ``g2`` (:math:`\bar D`) fixed, the orbital-Lagrangian response is
-
-        .. math::
-
-            (\dot{\bar A})^{m}_{p}
-            &=2(\dot F_{\mathrm{C}}+\dot{\bar F}_{\mathrm{A}})^{m}_{p},
-              &&m\in\mathbb C,\\
-            (\dot{\bar A})^{u}_{p}
-            &=\sum_v(\dot F_{\mathrm{C}})^{v}_{p}\bar\gamma^{v}_{u}
-              +\sum_{tvw}\dot{\langle pv|tw\rangle}\bar D_{tu,vw},
-              &&u\in\mathbb A,\\
-            (\dot{\bar A})^{e}_{p}&=0,
-              &&e\in\mathbb V.
-
-        The returned component for pair :math:`(p_I,q_I)` is therefore
-
-        .. math::
-
-            [\mathcal A^{\mathrm{oo}}\mathbf z]_I
-            =2\left[(\dot{\bar A})^{q_I}_{p_I}
-                    -(\dot{\bar A})^{p_I}_{q_I}\right].
-
-        This private kernel assumes that ``vector`` is real with shape
-        ``(nrot,)`` and that ``intermediates`` was built at the current orbitals
-        and stored ``g1``.  It performs no validation.  The AO Hamiltonian, DF
-        tensor, geometry, and both active RDMs are fixed; CI and nuclear response
-        are excluded.  The analytic, unscreened gradient expression is
-        differentiated.
-
-        Parameters
-        ----------
-        vector : np.ndarray
-            Real orbital-rotation direction in ``nrr`` C-order.
-        intermediates : tuple[np.ndarray, np.ndarray, np.ndarray]
-            ``(Fcore_mo, Fact_mo, B_ga)`` returned by
-            :meth:`_build_orbital_response_intermediates` at the same base point.
-
-        Returns
-        -------
-        np.ndarray
-            Fixed-RDM orbital Hessian applied to ``vector``, with shape
-            ``(nrot,)`` and the same pair ordering.
-        """
-        A_response = self._compute_orbital_lagrangian_response(vector, intermediates)
-        gradient_response = 2.0 * (A_response - A_response.T)
-        return self._mat_to_vec(gradient_response)
-
-    def _compute_orbital_hessian_transpose_vector_product(self, vector, intermediates):
-        r"""Evaluate the transpose fixed-RDM orbital Hessian action.
-
-        If :math:`H_{IJ}=D_J\bar g_I`, the derivative of the scalar
-        :math:`\mathbf z^{\mathsf T}\bar{\mathbf g}` with respect to pair
-        :math:`I` is
-
-        .. math::
-
-            [H^{\mathsf T}\mathbf z]_I
-            =2\left(B^{q_I}_{p_I}-B^{p_I}_{q_I}\right),
-            \qquad
-            B=\dot{\bar A}[\mathbf z]+Z\bar A-\bar A Z.
-
-        ``intermediates[3]`` contains the nonzero core and active columns of
-        the unsymmetrized base matrix :math:`\bar A`; its virtual columns are
-        zero and are not stored.  The commutator makes this application the exact
-        transpose of :meth:`_compute_orbital_hessian_vector_product`, even
-        when frozen rotations leave nonzero components of the full orbital
-        gradient.  No dense Hessian is formed.
-
-        Returns
-        -------
-        np.ndarray
-            :math:`H^{\mathsf T}\mathbf z` with shape ``(nrot,)``.
-        """
-        A_response = self._compute_orbital_lagrangian_response(vector, intermediates)
-        A_h = intermediates[3]
-        Z = self._vec_to_mat(vector)
-        holes = slice(self.core.start, self.actv.stop)
-        adjoint_A = A_response
-        adjoint_A[:, holes] += Z @ A_h
-        adjoint_A -= A_h @ Z[holes, :]
-        gradient_response = 2.0 * (adjoint_A - adjoint_A.T)
-        return self._mat_to_vec(gradient_response)
-
-    def _compute_orbital_lagrangian_response(self, vector, intermediates):
-        r"""Return the full MO Lagrangian response for an orbital direction.
-
-        This evaluates :math:`\dot{\bar A}[\mathbf z]` with the
-        state-averaged RDMs held fixed:
-
-        .. math::
-
-            \dot{\bar A}^{m}_{p}
-            &=2(\dot F_{\mathrm C}+\dot{\bar F}_{\mathrm A})^{m}_{p},
-              &&m\in\mathbb C,\\
-            \dot{\bar A}^{u}_{p}
-            &=\sum_v(\dot F_{\mathrm C})^{v}_{p}\bar\gamma^{v}_{u}
-              +\sum_{tvw}\dot{\langle pv|tw\rangle}\bar D_{tu,vw},
-              &&u\in\mathbb A,\\
-            \dot{\bar A}^{e}_{p}&=0,
-              &&e\in\mathbb V.
-
-        The Fock and integral responses are those documented by
-        :meth:`_compute_orbital_hessian_vector_product`.  The antisymmetric
-        part generates the orbital--orbital Hessian action,
-
-        .. math::
-
-            [\mathcal A^{\mathrm{oo}}\mathbf z]_I
-            =2\left[(\dot{\bar A})^{q_I}_{p_I}
-                    -(\dot{\bar A})^{p_I}_{q_I}\right].
-
-        The contribution to the relaxed orbital orthogonality multiplier
-        additionally contains the moving-frame commutator
-        :math:`Z\bar A-\bar A Z`; it is assembled by
-        :meth:`MCOptimizer.compute_omega`.
-
-        Parameters
-        ----------
-        vector : np.ndarray
-            Validated real orbital direction with shape ``(nrot,)``.
-        intermediates : tuple[np.ndarray, ...]
-            Base-point tensors from
-            :meth:`_build_orbital_response_intermediates`.
-
-        Returns
-        -------
-        np.ndarray
-            Full MO Lagrangian response matrix with shape ``(nmo, nmo)``.
-        """
-        Fcore_mo, Fact_mo, B_ga = intermediates[:3]
-        Z = self._vec_to_mat(vector)
-        C_response = self.C @ Z
-        Ccore_response = C_response[:, self.core]
-        Cact_response = C_response[:, self.actv]
-
-        Fcore_ao_response = self._build_transition_fock_response(
-            Ccore_response,
-            self.Ccore,
-            coulomb_factor=2.0,
-            exchange_factor=1.0,
-        )
-        Fact_ao_response = self._build_transition_fock_response(
-            Cact_response,
-            self.Cact,
-            self.g1,
-            coulomb_factor=1.0,
-            exchange_factor=0.5,
-        )
-
-        def transform_response(operator_mo, operator_ao_response):
-            r"""Transform an AO response while differentiating both MO legs.
-
-            .. math::
-
-                \dot F=Z^{\mathsf T}F+FZ+C^{\mathsf T}\dot F^{\mathrm{AO}}C.
-            """
-            return (
-                Z.T @ operator_mo
-                + operator_mo @ Z
-                + self._transform_ao_operator(operator_ao_response, self.C)
-            )
-
-        Fcore_response = transform_response(Fcore_mo, Fcore_ao_response)
-        Fact_response = transform_response(Fact_mo, Fact_ao_response)
-
-        B_response_ga = np.einsum("rp,Pru->Ppu", Z, B_ga, optimize=True)
-        B_response_ga += self._transform_df_block(self.C, Cact_response)
-        B_aa = B_ga[:, self.actv, :]
-        B_response_aa = B_response_ga[:, self.actv, :]
-
-        A_response = np.zeros_like(Fcore_response)
-        A_response[:, self.core] = 2.0 * (Fcore_response + Fact_response)[:, self.core]
-        A_response[:, self.actv] = np.einsum(
-            "rv,vu->ru", Fcore_response[:, self.actv], self.g1, optimize=True
-        )
-        contracted_base = np.einsum("Pvw,tuvw->Ptu", B_aa, self.g2, optimize=True)
-        contracted_response = np.einsum(
-            "Pvw,tuvw->Ptu", B_response_aa, self.g2, optimize=True
-        )
-        A_response[:, self.actv] += np.einsum(
-            "Prt,Ptu->ru", B_response_ga, contracted_base, optimize=True
-        )
-        A_response[:, self.actv] += np.einsum(
-            "Prt,Ptu->ru", B_ga, contracted_response, optimize=True
-        )
-        return A_response
-
-    def _build_ci_orbital_response_intermediates(self):
-        r"""Build fixed tensors for the orbital--CI response block.
-
-        Returns the current inactive-core Fock matrix and transformed DF block
-
-        .. math::
-
-            F_{\mathrm C}=C^{\mathsf T}F_{\mathrm C}^{\mathrm{AO}}C,\qquad
-            B^P_{ru}=\sum_{\mu\nu}C_{\mu r}B^P_{\mu\nu}C_{\nu u}.
-
-        Both are independent of a CI multiplier direction and can be reused
-        for every column of the orbital--CI block.
-
-        Returns
-        -------
-        tuple[np.ndarray, np.ndarray]
-            The pair (Fcore_mo, B_ga), with shapes (nmo, nmo) and
-            (naux, nmo, nact), respectively.
-        """
-        Fcore_ao = self.fock_builder.build_core_fock(self.Ccore, hcore=self.hcore)
-        Fcore_mo = self._transform_ao_operator(Fcore_ao, self.C)
-        B_ga = self._transform_df_block(self.C, self.Cact)
-        return Fcore_mo, B_ga
-
-    def _build_orbital_ci_response_intermediates(self):
-        r"""Build fixed tensors for the orbital contribution to the CI row.
-
-        For the current real restricted orbitals, this routine returns
-
-        .. math::
-
-            F_{\mathrm C}^{\mathrm{AO}}
-            &=h+2J[P_{\mathrm C}]-K[P_{\mathrm C}],\\
-            F_{\mathrm C}&=C^{\mathsf T}F_{\mathrm C}^{\mathrm{AO}}C,\\
-            B^P_{pu}&=\sum_{\mu\nu}C_{\mu p}B^P_{\mu\nu}C_{\nu u}.
-
-        They are the base-point tensors needed to differentiate the scalar,
-        one-electron, and two-electron parts of the active-space Hamiltonian
-        along any orbital direction.  The returned tensors contain no
-        response themselves and remain valid only while the orbitals, AO
-        Hamiltonian, and DF factors are unchanged.
-
-        Returns
-        -------
-        tuple[np.ndarray, np.ndarray, np.ndarray]
-            ``(Fcore_ao, Fcore_mo, B_ga)`` with shapes ``(nao, nao)``,
-            ``(nmo, nmo)``, and ``(naux, nmo, nact)``, respectively.
-        """
-        Fcore_ao = self.fock_builder.build_core_fock(self.Ccore, hcore=self.hcore)
-        Fcore_mo = self._transform_ao_operator(Fcore_ao, self.C)
-        B_ga = self._transform_df_block(self.C, self.Cact)
-        return Fcore_ao, Fcore_mo, B_ga
-
-    def _compute_active_space_hamiltonian_response(self, vector, intermediates):
-        r"""Differentiate the active-space Hamiltonian along an orbital vector.
-
-        With ``vector`` embedded in the antisymmetric generator :math:`Z` by
-        :meth:`_vec_to_mat`, let :math:`\dot C=CZ`.  This kernel returns the
-        integral triple defining the response Hamiltonian
-
-        .. math::
-
-            \hat H[\mathbf z]
-            =\dot E_{\mathrm C}
-             +\sum_{uv}(\dot F_{\mathrm C})^{u}_{v}\hat E^{u}_{v}
-             +\frac12\sum_{uvtw}
-              \dot{\langle uv|tw\rangle}\hat E^{uv}_{tw}.
-
-        Here
-
-        .. math::
-
-            \dot F_{\mathrm C}^{\mathrm{AO}}
-            &=2J[\dot P_{\mathrm C}]-K[\dot P_{\mathrm C}],\\
-            \dot F_{\mathrm C}
-            &=Z^{\mathsf T}F_{\mathrm C}+F_{\mathrm C}Z
-              +C^{\mathsf T}\dot F_{\mathrm C}^{\mathrm{AO}}C,
-
-        and :math:`\dot E_{\mathrm C}` is the derivative of
-        :math:`\mathrm{Tr}[C_{\mathrm C}^{\mathsf T}
-        (h+F_{\mathrm C}^{\mathrm{AO}})
-        C_{\mathrm C}]`.  The two-electron response differentiates all four
-        active indices of
-        :math:`\langle uv|tw\rangle=\sum_P B^P_{ut}B^P_{vw}`.
-
-        This private kernel assumes a validated real vector with shape
-        ``(nrot,)`` and intermediates from
-        :meth:`_build_orbital_ci_response_intermediates` at the same base
-        point.  The AO integrals and nuclear geometry are fixed.
-
-        Returns
-        -------
-        tuple[float, np.ndarray, np.ndarray]
-            Scalar core-energy response, active one-electron response, and
-            active two-electron response in the CI solver's physicists'
-            convention.
-        """
-        Fcore_ao, Fcore_mo, B_ga = intermediates
-        Z = self._vec_to_mat(vector)
-        C_response = self.C @ Z
-        Ccore_response = C_response[:, self.core]
-
-        Fcore_ao_response = self._build_transition_fock_response(
-            Ccore_response,
-            self.Ccore,
-            coulomb_factor=2.0,
-            exchange_factor=1.0,
-        )
-        Fcore_response = (
-            Z.T @ Fcore_mo
-            + Fcore_mo @ Z
-            + self._transform_ao_operator(Fcore_ao_response, self.C)
-        )
-
-        h_plus_fcore = self.hcore + Fcore_ao
-        scalar_response = np.trace(
-            Ccore_response.T @ h_plus_fcore @ self.Ccore
-            + self.Ccore.T @ h_plus_fcore @ Ccore_response
-            + self.Ccore.T @ Fcore_ao_response @ self.Ccore
-        )
-
-        B_active = B_ga[:, self.actv, :]
-        B_left_response = np.einsum("ru,Prv->Puv", Z[:, self.actv], B_ga, optimize=True)
-        B_active_response = B_left_response + B_left_response.transpose(0, 2, 1)
-        eri_response = np.einsum(
-            "Put,Pvw->uvtw", B_active_response, B_active, optimize=True
-        )
-        eri_response += np.einsum(
-            "Put,Pvw->uvtw", B_active, B_active_response, optimize=True
-        )
-
-        one_body_response = Fcore_response[self.actv, self.actv]
-        return (
-            float(scalar_response),
-            np.ascontiguousarray(one_body_response),
-            np.ascontiguousarray(eri_response),
-        )
-
-    def _build_orbital_lagrangian_from_rdms(
-        self,
-        overlap_response,
-        g1_response,
-        g2_response,
-        intermediates,
-    ):
-        r"""Build a full MO Lagrangian matrix from overlap and active RDMs.
-
-        ``overlap_response`` supplies the scalar :math:`s`, while
-        ``g1_response`` and ``g2_response`` supply :math:`\gamma` and the
-        spin-free two-particle density used to form :math:`D`.  The returned
-        unsymmetrized matrix is
-
-        .. math::
-
-            A^{m}_{p}[s,\gamma,D]
-            &=2\left[s(F_{\mathrm C})^{m}_{p}
-                     +(F_{\mathrm A}[\gamma])^{m}_{p}\right],
-              &&m\in\mathbb C,\\
-            A^{u}_{p}[s,\gamma,D]
-            &=\sum_v(F_{\mathrm C})^{v}_{p}\gamma^{v}_{u}
-              +\sum_{tvw}\langle pv|tw\rangle D_{tu,vw},
-              &&u\in\mathbb A,\\
-            A^{e}_{p}[s,\gamma,D]&=0,
-              &&e\in\mathbb V.
-
-        For a normalized target state the inputs are
-        :math:`(1,\gamma^\alpha,\Gamma^\alpha)`.  For a CI multiplier they
-        are the root-summed bra-plus-ket transition quantities
-        :math:`(s[\mathbf x],\gamma[\mathbf x],\Gamma[\mathbf x])`.
-
-        Returns
-        -------
-        np.ndarray
-            Full unsymmetrized MO Lagrangian matrix with shape ``(nmo, nmo)``.
-        """
-        Fcore_mo, B_ga = intermediates
-        Fact_ao_response = self._build_transition_fock_response(
-            self.Cact,
-            self.Cact,
-            0.5 * g1_response,
-            coulomb_factor=1.0,
-            exchange_factor=0.5,
-        )
-        Fact_response = self._transform_ao_operator(Fact_ao_response, self.C)
-        g2_working_response = self._make_working_2rdm(g2_response)
-
-        A_response = np.zeros_like(Fcore_mo)
-        A_response[:, self.core] = (
-            2.0 * (overlap_response * Fcore_mo + Fact_response)[:, self.core]
-        )
-        A_response[:, self.actv] = np.einsum(
-            "rv,vu->ru", Fcore_mo[:, self.actv], g1_response, optimize=True
-        )
-        B_aa = B_ga[:, self.actv, :]
-        contracted_density = np.einsum(
-            "Pvw,tuvw->Ptu", B_aa, g2_working_response, optimize=True
-        )
-        A_response[:, self.actv] += np.einsum(
-            "Prt,Ptu->ru", B_ga, contracted_density, optimize=True
-        )
-        return A_response
-
-    def _compute_ci_orbital_response_from_rdms(
-        self,
-        overlap_response,
-        g1_response,
-        g2_response,
-        intermediates,
-    ):
-        r"""Map symmetrized transition RDMs to the orbital equation.
-
-        For a root-summed CI multiplier direction, overlap_response is s[x],
-        while g1_response and g2_response are the raw spin-free, bra-plus-ket
-        transition RDMs.  This routine evaluates
-
-        .. math::
-
-            (A^{\mathrm{oc}}[\mathbf x])^{m}_{p}
-            &=2\left[s[\mathbf x](F_{\mathrm C})^{m}_{p}
-                    +(F_{\mathrm A}[\gamma[\mathbf x]])^{m}_{p}\right],\\
-            (A^{\mathrm{oc}}[\mathbf x])^{u}_{p}
-            &=\sum_v(F_{\mathrm C})^{v}_{p}\gamma^{v}_{u}[\mathbf x]
-              +\sum_{tvw}\langle pv|tw\rangle D_{tu,vw}[\mathbf x],\\
-            [\mathcal A^{\mathrm{oc}}\mathbf x]_I
-            &=2\left[(A^{\mathrm{oc}})^{q_I}_{p_I}
-                    -(A^{\mathrm{oc}})^{p_I}_{q_I}\right].
-
-        The first equation applies to a core column i, the second to an active
-        column u, and virtual columns are zero.  The working density D[x] is
-        formed from g2_response using the same permutation and symmetrization
-        as set_rdms.
-
-        Parameters
-        ----------
-        overlap_response : float
-            Root-summed bra-plus-ket transition overlap.
-        g1_response : np.ndarray
-            Root-summed spin-free transition 1-RDM, shape (nact, nact).
-        g2_response : np.ndarray
-            Root-summed spin-free transition 2-RDM, shape
-            (nact, nact, nact, nact).
-        intermediates : tuple[np.ndarray, np.ndarray]
-            The pair (Fcore_mo, B_ga) returned by
-            _build_ci_orbital_response_intermediates at the same base point.
-
-        Returns
-        -------
-        np.ndarray
-            Orbital part of the orbital--CI action, with shape (nrot,) and
-            nrr C-order.
-        """
-        A_response = self._build_orbital_lagrangian_from_rdms(
-            overlap_response,
-            g1_response,
-            g2_response,
-            intermediates,
-        )
-        gradient_response = 2.0 * (A_response - A_response.T)
-        return self._mat_to_vec(gradient_response)
 
     def _compute_reference_energy(self):
         energy = self.Ecore + self.e_nuc
